requirements.txt
# Core dependencies
numpy>=1.19.2
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.3
Pillow>=8.3.1
matplotlib>=3.4.3
scikit-learn>=0.24.2

# Segmentation dependencies
segmentation-models-pytorch>=0.2.0
albumentations>=1.0.3

# Utilities
tqdm>=4.62.2
pyyaml>=5.4.1
pandas>=1.3.3

# Optional geospatial dependencies (install separately if needed)
# gdal>=3.2.2
# fiona>=1.8.20
# geopandas>=0.9.0
# shapely>=1.7.1
